pyproject.toml
[project]
name = "chatbot-agent-resource"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gradio>=6.10.0",
    "ollama>=0.4.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
]
